requirements.txt
openpyxl
selenium
mysql.connector